AI/requirements.txt
fastapi
uvicorn
google-generativeai
opencv-python
pydantic
python-dotenv
pydantic-settings>=2.0.0
